Rename new geometry information sample to avoid windows path length limitations (#1821)
