+++ src/MAUI/Maui.Samples/Samples/Geometry/DisplayGeometryEditorInteractionInformation/DisplayGeometryEditorInteractionInformation.xaml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage x:Class="ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction.DisplayGeometryEditorInformationDuringInteraction"
+<ContentPage x:Class="ArcGIS.Samples.DisplayGeometryEditorInteractionInformation.DisplayGeometryEditorInteractionInformation"
              xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:converters="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
