+++ src/MAUI/Maui.Samples/Samples/Geometry/DisplayGeometryEditorInteractionInformation/displaygeometryeditorinteractioninformation.jpg
