+++ src/MAUI/Maui.Samples/Samples/Geometry/DisplayGeometryEditorInteractionInformation/readme.md
@@ -2,7 +2,7 @@
 
 Use the geometry editor to see information about the geometry editor's previewed geometry during an editing interaction.
 
-![DisplayGeometryEditorInformationDuringInteraction](displaygeometryeditorinformationduringinteraction.jpg)
+![DisplayGeometryEditorInteractionInformation](displaygeometryeditorinteractioninformation.jpg)
 
 ## Use case
