+++ src/Samples.Shared/Resources/FeaturedSamples.xml
@@ -17,7 +17,7 @@
 		</Category>
 		<Category name="Geometry">
 			<Sample>CreateAndEditGeometries</Sample>
-			<Sample>DisplayGeometryEditorInformationDuringInteraction</Sample>
+			<Sample>DisplayGeometryEditorInteractionInformation</Sample>
 			<Sample>EditGeometriesWithProgrammaticReticleTool</Sample>
 			<Sample>SnapGeometryEdits</Sample>
 		</Category>
