+++ src/WPF/WPF.Viewer/Samples/Geometry/DisplayGeometryEditorInteractionInformation/DisplayGeometryEditorInteractionInformation.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.DisplayGeometryEditorInformationDuringInteraction.DisplayGeometryEditorInformationDuringInteraction"
+﻿<UserControl x:Class="ArcGIS.WPF.Samples.DisplayGeometryEditorInteractionInformation.DisplayGeometryEditorInteractionInformation"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
