+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Geometry/DisplayGeometryEditorInteractionInformation/DisplayGeometryEditorInteractionInformation.jpg
