+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Geometry/DisplayGeometryEditorInteractionInformation/DisplayGeometryEditorInteractionInformation.xaml
@@ -1,5 +1,5 @@
 <UserControl 
-    x:Class="ArcGIS.WinUI.Samples.DisplayGeometryEditorInformationDuringInteraction.DisplayGeometryEditorInformationDuringInteraction"
+    x:Class="ArcGIS.WinUI.Samples.DisplayGeometryEditorInteractionInformation.DisplayGeometryEditorInteractionInformation"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esri="using:Esri.ArcGISRuntime"
