+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Geometry/DisplayGeometryEditorInteractionInformation/DisplayGeometryEditorInteractionInformation.xaml.cs
@@ -1,21 +1,24 @@
-﻿// Copyright 2026 Esri.
+// Copyright 2026 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an 
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
 // language governing permissions and limitations under the License.
 
 using Esri.ArcGISRuntime.Geometry;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Symbology;
 using Esri.ArcGISRuntime.UI;
+using Esri.ArcGISRuntime.UI.Controls;
 using Esri.ArcGISRuntime.UI.Editing;
-using Microsoft.Maui.ApplicationModel;
-using Map = Esri.ArcGISRuntime.Mapping.Map;
+using Microsoft.UI.Xaml;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
 
-namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
+namespace ArcGIS.WinUI.Samples.DisplayGeometryEditorInteractionInformation
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
         name: "Display geometry editor information during interaction",
@@ -23,8 +26,9 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
         description: "Use the geometry editor to see information about the geometry editor's previewed geometry during an editing interaction.",
         instructions: "Tap a graphic to edit its geometry by moving, rotating, or scaling the geometry. During the interaction, information about the geometry will be displayed to provide feedback to the user.",
         tags: new[] { "draw", "edit", "geometry editor", "interaction preview" })]
-    public partial class DisplayGeometryEditorInformationDuringInteraction
+    public partial class DisplayGeometryEditorInteractionInformation
     {
+
         // Create a geometry editor instance.
         private readonly GeometryEditor _geometryEditor = new();
 
@@ -41,7 +45,7 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
 
         private readonly Multipoint _redlandsMultipoint = Multipoint.FromJson(@"{""points"":[[-13045283.292102993,4035739.1925106063],[-13045314.922186911,4036533.8852012255],[-13044798.24723932,4036138.7808295386],[-13044354.514637273,4035719.3623426706],[-13044281.57229173,4036473.0999132735]],""spatialReference"":{""wkid"":3857}}") as Multipoint;
 
-        public DisplayGeometryEditorInformationDuringInteraction()
+        public DisplayGeometryEditorInteractionInformation()
         {
             InitializeComponent();
             _ = Initialize();
@@ -111,7 +115,7 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
                 // Update the UI labels with the scale factor information.
                 InteractionPreviewDescriptionLabel.Text = "Scale Factor (X, Y):";
                 InteractionPreviewValueLabel.Text = $"({scaleX:F2}, {scaleY:F2})";
-                InteractionPreviewGrid.IsVisible = true;
+                InteractionPreviewGrid.Visibility = Visibility.Visible;
             }
         }
 
@@ -156,7 +160,7 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
 
             // Update the UI label for rotation angle.
             InteractionPreviewDescriptionLabel.Text = "Rotation Angle (degrees):";
-            InteractionPreviewGrid.IsVisible = true;
+            InteractionPreviewGrid.Visibility = Visibility.Visible;
         }
 
         // Set the moving geometry label based on the interaction preview.
@@ -168,7 +172,7 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
             // Update the UI labels with the center point information.
             InteractionPreviewDescriptionLabel.Text = "Center (X, Y):";
             InteractionPreviewValueLabel.Text = $"({previewCenter.X:F2}, {previewCenter.Y:F2})";
-            InteractionPreviewGrid.IsVisible = true;
+            InteractionPreviewGrid.Visibility = Visibility.Visible;
         }
 
         // Handle the InteractionPreviewChanged event to update the UI based on the type of interaction.
@@ -181,25 +185,25 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
                 switch (e.InteractionPreview.InteractionType)
                 {
                     case GeometryEditorInteractionType.Scale:
-                        MainThread.BeginInvokeOnMainThread(() => SetScaleFactorLabel(e.InteractionPreview));
+                        DispatcherQueue.TryEnqueue(() => SetScaleFactorLabel(e.InteractionPreview));
                         break;
                     case GeometryEditorInteractionType.Rotate:
-                        MainThread.BeginInvokeOnMainThread(() => SetRotationAngleLabel(e.InteractionPreview));
+                        DispatcherQueue.TryEnqueue(() => SetRotationAngleLabel(e.InteractionPreview));
                         break;
                     case GeometryEditorInteractionType.Move:
-                        MainThread.BeginInvokeOnMainThread(() => SetMovingGeometryLabel(e.InteractionPreview));
+                        DispatcherQueue.TryEnqueue(() => SetMovingGeometryLabel(e.InteractionPreview));
                         break;
                 }
             }
             else
             {
                 // If the interaction preview is null the interaction has finished.
-                MainThread.BeginInvokeOnMainThread(() => InteractionPreviewGrid.IsVisible = false);
+                DispatcherQueue.TryEnqueue(() => InteractionPreviewGrid.Visibility = Visibility.Collapsed);
             }
         }
 
         #region UI Event Handlers
-        private async void MyMapView_GeoViewTapped(object sender, Esri.ArcGISRuntime.Maui.GeoViewInputEventArgs e)
+        private async void MyMapView_GeoViewTapped(object sender, GeoViewInputEventArgs e)
         {
             if (!_geometryEditor.IsStarted)
             {
@@ -217,32 +221,32 @@ namespace ArcGIS.Samples.DisplayGeometryEditorInformationDuringInteraction
             }
         }
 
-        private void SaveButton_Click(object sender, EventArgs e)
+        private void SaveButton_Click(object sender, RoutedEventArgs e)
         {
             if (_geometryEditor.IsStarted)
             {
                 _editingGraphic.Geometry = _geometryEditor.Stop();
                 _editingGraphic.IsVisible = true; // Show the graphic again after editing.
-                InteractionPreviewGrid.IsVisible = false;
+                InteractionPreviewGrid.Visibility = Visibility.Collapsed;
             }
         }
 
-        private void CancelButton_Click(object sender, EventArgs e)
+        private void CancelButton_Click(object sender, RoutedEventArgs e)
         {
             if (_geometryEditor.IsStarted)
             {
                 _geometryEditor.Stop();
                 _editingGraphic.IsVisible = true; // Show the graphic again after canceling.
-                InteractionPreviewGrid.IsVisible = false;
+                InteractionPreviewGrid.Visibility = Visibility.Collapsed;
             }
         }
 
-        private void RedoButton_Click(object sender, EventArgs e)
+        private void RedoButton_Click(object sender, RoutedEventArgs e)
         {
             _geometryEditor.Redo();
         }
 
-        private void UndoButton_Click(object sender, EventArgs e)
+        private void UndoButton_Click(object sender, RoutedEventArgs e)
         {
             _geometryEditor.Undo();
         }
